requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==1.26.4
regex==2024.7.24
# Optional (comment out if unused)
openai==1.37.0
anthropic==0.34.2
# For vector memory (optional)
faiss-cpu==1.8.0.post1
langchain
langchain-openai
